Moved test_feedwriter fixtures to conftest.py
